refactored added implementation to main class (application)

+++ src/main/java/our/project/model/HasFieldForEvenSort.java
@@ -1,5 +1,5 @@
 package our.project.model;
 
 public interface HasFieldForEvenSort {
-    int getFieldForEvenSort();
+    long getFieldForEvenSort();
 }
